qgrep can now be compiled for Linux.
This change fixes various compilation errors and adds POSIX implementations for fileutil and strprintf, as well as minor refactoring (i.e. strprintf is now a part of stringutil). Code now compiles and almost works with GCC 4.6.3.

+++ src/streamutil.hpp
@@ -5,7 +5,7 @@
 inline bool read(std::istream& in, void* data, size_t size)
 {
 	in.read(static_cast<char*>(data), size);
-	return in.gcount() == size;
+	return static_cast<size_t>(in.gcount()) == size;
 }
 
 template <typename T> inline bool read(std::istream& in, T& value)
